requirements.txt
        certifi==2024.2.2
        charset-normalizer==3.3.2
        click==8.1.7
        Flask==3.0.2
        idna==3.6
        Jinja2==3.1.3
        MarkupSafe==2.1.5
        numpy==1.26.4
        Pillow==10.2.0
        requests==2.31.0
        torch==2.3.0
        torchvision==0.18.0
        urllib3==2.2.1
        Werkzeug==3.1.0
        
